pyproject.toml
[project]
name = "doc-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
